add button to project selection dialog
